refactor(e18e): split lib.rs into scanner, helpers, and per-topic rule modules
The 2570-line crates/e18e/src/lib.rs was split into 11 focused modules so no
file is far from the ~350-line target. The original Scanner struct + impl is
split across additional impl blocks in each rule module, following the same
pattern already used by the stylistic crate. Public API (RULE_NAMES, scan_e18e,
Diagnostic types) is unchanged. All existing unit tests still pass.
